requirements.txt
# =====================================================
# CORE FLASK & WEB FRAMEWORK
# =====================================================
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7

# =====================================================
# DATABASE & ORM
# =====================================================
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
alembic==1.13.0

# =====================================================
# AUTHENTICATION & SECURITY
# =====================================================
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
PyJWT==2.8.0
bcrypt==4.1.1
cryptography==41.0.7

# =====================================================
# API & CORS
# =====================================================
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
flask-restx==1.3.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# =====================================================
# VALIDATION & SERIALIZATION
# =====================================================
python-dotenv==1.0.0
email-validator==2.1.0
phonenumbers==8.13.26
pydantic==2.5.2

# =====================================================
# DATE & TIME
# =====================================================
python-dateutil==2.8.2
pytz==2023.3
arrow==1.3.0

# =====================================================
# EMAIL
# =====================================================
Flask-Mail==0.9.1
sendgrid==6.11.0

# =====================================================
# FILE STORAGE (Cloudinary / AWS S3)
# =====================================================
cloudinary==1.36.0
boto3==1.34.14

# =====================================================
# CACHE & SESSIONS
# =====================================================
Flask-Caching==2.1.0
Flask-Session==0.5.0
# redis==5.0.1 # Removed to save space, assuming simple cache fallback

# =====================================================
# RATE LIMITING
# =====================================================
Flask-Limiter==3.5.0

# =====================================================
# MONITORING & LOGGING
# =====================================================
sentry-sdk[flask]==1.39.1
python-json-logger==2.0.7

# =====================================================
# UTILITIES
# =====================================================
requests==2.31.0
urllib3==2.1.0
python-slugify==8.0.1
bleach==6.1.0
shortuuid==1.0.11

# =====================================================
# WSGI SERVER
# =====================================================
gunicorn==21.2.0
